+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 ### [@project-version@](https://github.com/wow-addon-dev/Library/tree/@project-version@) (@build-date@)
 [Full Changelog](@full-changelog@) - [Previous Releases](https://github.com/wow-addon-dev/Library/releases)
 
-- Added: TOC version for patch 5.5.4 [mist of pandaria - classic]
-- Removed: TOC version for patch 12.0.1 [retail]
+- Added: Optional cancel callback support for confirmation dialogs
+- Added: TOC version for patch 12.0.7 [retail]
+